fix: change moduleName -> mockModuleName
```
the module factory of `jest.mock()` is not allowed to reference any out-of-scope variables.     Invalid variable access: moduleName
```

This fixes #216 for me using TypeScript using @latest for all libraries (i.e. next-11.0.2) in a TypeScript environment.

There would seem to be something "lost in translation" between the `.ts -> .js with types.d.ts` conversion for the npm package.

This solution comes from a monkey-patch, but it's a simple refactor of one variable name on 4 lines, so it should be good to go.